requirements.txt
Django==4.2.14
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
dj-database-url==2.2.0
whitenoise==6.7.0
pytest==8.3.2
pytest-django==4.8.0
python-dotenv==1.0.1
